[TASK-3] fix: export CurrencyValue type and fix decimal parsing

+++ victoria.glushkova/task-3/internal/currency/currency.go
@@ -19,7 +19,7 @@ func ConvertAndSort(valCurs *xml.ValCurs) []Currency {
 		currencies[i] = Currency{
 			NumCode:  valute.NumCode,
 			CharCode: valute.CharCode,
-			Value:    valute.Value,
+			Value:    float64(valute.Value),
 		}
 	}
